Okay. So here's the deal (in description):
Well, after some testing, at protocol 91 and below, both client and server said outdated server. That means that 91 is definitely outdated. From 92 and above, only the server said outdated, the client tried to connect. This means that 92 is the correct protocol, but there is simply a lack of packets. The next step is to find out which packet is missing, but for now, we know the protocol, so DON'T CHANGE IT!